add: procedures, however, they vanish ONLY after recompile. and I dont know why add: inserts and selects from usos
