+++ Integration/Query/Queries.py
@@ -1,10 +1,14 @@
 class Queries:
-    
-    selectCliente = "SELECT * FROM cliente;"
-    selectPrato = "SELECT * FROM prato;"
-    selectFornecedor = "SELECT * FROM fornecedor;"
-    selectIngredientes = "SELECT * FROM ingredientes;"
-    selectVenda = "SELECT * FROM venda;"
 
     def chooseSelect(table):
         return f"SELECT * FROM {table};"
+    
+    callSorteio = "CALL Sorteio();"
+    callEstatisticas = "CALL Estatisticas_Venda()"
+    
+    def callReajuste(numDecimal):
+        return f"CALL Reajuste({numDecimal})"
+    
+    def callGastarPontos(idCliente, idPrato):
+        return f"CALL Reajuste({idCliente}, {idPrato})"
+
